requirements.txt
Flask==2.0.3
torch==1.13.1
numpy==1.21.6
Pillow==9.3.0
scikit-image==0.18.3
opencv-python-headless==4.5.4.60
